fix: user dao

+++ src/daos/user.dao.js
@@ -9,6 +9,19 @@ export class UserDAO {
     return User.findById(id);
   }
 
+  async findByIdAndActive(id) {
+    const user = await this.findById(id);
+
+    if (!user) {
+      throw new Error('User not found');
+    }
+
+    if (user.deletedAt === null && user.bannedAt === null) {
+      return user;
+    }
+    throw new Error('User is deleted or banned');
+  }
+
   async findAll(filter = {}) {
     return User.find(filter);
   }
@@ -19,14 +32,26 @@ export class UserDAO {
   }
 
   async updateById(userId, data) {
+    const isActive = await this.isActive(userId);
+    if (!isActive) {
+      throw new Error('User is deleted or banned');
+    }
     return User.findByIdAndUpdate(userId, data, { new: true });
   }
 
   async delete(userId) {
+    const isActive = await this.isActive(userId);
+    if (!isActive) {
+      throw new Error('User is deleted or banned');
+    }
     return User.findByIdAndDelete(userId);
   }
 
   async updateOtp(email, otpData) {
+    const isActive = await this.isActive(email);
+    if (!isActive) {
+      throw new Error('User is deleted or banned');
+    }
     return User.findOneAndUpdate(
       { email },
       { $push: { OTP: otpData } },
@@ -34,15 +59,11 @@ export class UserDAO {
     );
   }
 
-  async updatePassword(userId, newPassword) {
-    return User.findByIdAndUpdate(
-      userId,
-      { password: newPassword },
-      { new: true },
-    );
-  }
-
   async updateRefreshToken(userId, token) {
+    const isActive = await this.isActive(userId);
+    if (!isActive) {
+      throw new Error('User is deleted or banned');
+    }
     return User.findByIdAndUpdate(
       userId,
       { refreshToken: token },
@@ -51,17 +72,27 @@ export class UserDAO {
   }
 
   async isActive(userId) {
-    const user = await this.findById(userId);
-    return user.deletedAt === null;
+    const user =
+      (await this.findById(userId)) || (await this.findByEmail(userId));
+    if (user.deletedAt === null && user.bannedAt === null) {
+      return true;
+    }
+    return false;
   }
 
   async isDeleted(userId) {
     const user = await this.findById(userId);
-    return user.deletedAt !== null;
+    if (user.deletedAt !== null) {
+      return false;
+    }
+    return true;
   }
 
   async isBanned(userId) {
     const user = await this.findById(userId);
-    return user.bannedAt !== null;
+    if (user.bannedAt !== null) {
+      return false;
+    }
+    return true;
   }
 }
